Module faucon_asm::operands

Representations of Falcon instruction operands.

Structs

Register

A Falcon CPU register.

Enums

MemoryAccess

A direct memory access to an address in a MemorySpace.

MemorySpace

The used Falcon memory spaces in SRAM.

Operand

An operand of a Falcon assembly crate::Instruction.

RegisterKind

Types of CPU registers supported by the Falcon.

Functions

get_flag_name

Gets the qualified name of a flag bit in the $flags register based on the given bit index.

get_spr_name

Gets the qualified name of a special-purpose register based on the given register index.