Enum faucon_asm::isa::InstructionKind

pub enum InstructionKind {
    CMPU,
    CMPS,
    CMP,
    ADD,
    ADC,
    SUB,
    SBB,
    SHL,
    SHR,
    SAR,
    SHLC,
    SHRC,
    SETHI,
    CLEAR,
    MULU,
    MULS,
    SEXT,
    AND,
    OR,
    XOR,
    XBIT,
    BSET,
    BCLR,
    BTGL,
    DIV,
    MOD,
    SETP,
    MOV,
    LD,
    ST,
    PUSH,
    POP,
    CALL,
    LCALL,
    LJMP,
    RET,
    EXIT,
    SLEEP,
    PTLB,
    VTLB,
    ITLB,
    IRET,
    TRAP,
    XCLD,
    XDLD,
    XDST,
    XCWAIT,
    XDWAIT,
    IOWR,
    IOWRS,
    IORD,
    XXX,
}

Assembly instruction kinds within the Falcon ISA.

Through internal implementation details, this enum is responsible for generating opcode lookup tables that can be used to identify instructions and their variants.

Variants

CMPU

The CMPU instruction.

Compares two unsigned values and sets ALU flags based on the result.

CMPS

The CMPS instruction.

Compares two signed values and sets ALU flags based on the result.

CMP

The CMP instruction.

Compares two values and sets ALU flags based on the result.

ADD

The ADD instruction.

Computes the sum of two operands and stores the result.

ADC

The ADC instruction.

Computes the sum of two operands with a carry and stores the result.

SUB

The SUB instruction.

Subtracts two operands and stores the result.

SBB

The SBB instruction.

Subtracts two operands with borrow and stores the result.

SHL

The SHL instruction.

Shifts a value left and stores the result.

SHR

The SHR instruction.

Shifts a value right and stores the result.

SAR

The SAR instruction.

Shifts a value right with sign bit and stores the result.

SHLC

The SHLC instruction.

Shifts a value left with carry in and stores the result.

SHRC

The SHRC instruction.

Shifts a value right with carry in and stores the result.

SETHI

The SETHI instruction.

Sets the high 16 bits of a register to a value, without thouching the low 16 bits.

CLEAR

The CLEAR instruction.

Clears the contents of a register.

MULU

THE MULU instruction.

Performs an unsigned multiplication and stores the result.

MULS

The MULS instruction.

Performs a signed multiplication and stores the result.

SEXT

The SEXT instruction.

Sign-extends a value and stores the result.

AND

The AND instruction.

Performs a binary AND operation on two operands.

OR

The OR instruction.

Performs a binary OR operation on two operands.

XOR

The XOR instruction.

Performs a binary XOR operation on two operands.

XBIT

The XBIT instruction.

Extracts a bit from a specified register and stores it in the lowest bit of the destination register, setting all other bits to 0.

BSET

The BSET instruction.

Sets a specific bit in a given register.

BCLR

The BCLR instruction.

Clears a specific bit in a given register.

BTGL

The BTGL instruction.

Toggles (flips) a specific bit in a given register.

DIV

The DIV instruction.

Performs unsigned 32-bit division on two operands.

MOD

The MOD instruction.

Takes the modulus of two 32-bit unsigned operands.

SETP

The SETP instruction.

Sets a given bit in the $flags register to the lowest bit of the source register.

MOV

The MOV instruction.

Moves values of immediates or registers to other registers.

LD

The LD instruction.

Loads a value from Falcon DMem to a register.

ST

The ST instruction.

Stores a value from a register to Falcon DMem.

PUSH

The PUSH instruction.

Pushes a value onto the stack and increments the stack pointer by four.

POP

THE POP instruction.

Pops a value off the stack and increments the stack pointer by four.

CALL

The CALL instruction.

Performs an unconditional branch to an absolute address, pushing the return address onto the stack.

LCALL

The LCALL instruction.

Performs an unconditional branch to an absolute address, pushing the return address onto the stack.

LJMP

The LJMP instruction.

Performs an unconditional branch to an absolute address.

RET

The RET instruction.

Returns from a previous subroutine call.

EXIT

The EXIT instruction.

Halts microcode execution and triggers the EXIT interrupt so that the processor can only be restarted by the host machine.

SLEEP

The SLEEP instruction.

Puts the processor into sleep state until an unmasked interrupt is received. Repeated until the given flag bit is cleared.

PTLB

The PTLB instruction.

Loads the TLB that covers a given physical page into a destination register.

VTLB

The VTLB instruction.

Loads the TLB that covers a given virtual address into a destination register.

ITLB

The ITLB instruction.

Clears a non-secret TLB entry corresponding to a specified physical page.

IRET

The IRET instruction.

Returns from an interrupt handler.

TRAP

The TRAP instruction.

Triggers a software trap.

XCLD

The XCLD instruction.

Submits a DMA transfer request to load code from external memory.

XDLD

The XDLD instruction.

Submits a DMA transfer request to load data from external memory.

XDST

The XDST instruction.

Submits a DMA transfer request to store local Falcon data in external memory.

XCWAIT

The XCWAIT instruction.

Waits for all DMA code load transfers to complete.

XDWAIT

The XDWAIT instruction.

Waits for all DMA data load/store transfers to complete.

IOWR

The IOWR instruction.

Asynchronously writes a word to the I/O space of the microprocessor.

IOWRS

The IOWRS instruction.

Synchronously writes a word to the I/O space of the microprocessor.

IORD

The IORD instruction.

Reads a word from the I/O space of the processor.

XXX

An invalid or unknown instruction.

Implementations

impl InstructionKind

pub fn invalid(&self) -> bool

Checks if the instruction is invalid or unknown.

pub fn parse_sized_form_1(
a: u8,
b: u8,
subopcode: u8
) -> Option<InstructionMeta>

Parses a sized instruction in form 1.

This covers the opcode range from 0x00 to 0xBF. Form 1 essentially means that a decides the form and b selects the subopcode.

pub fn parse_sized_form_2(b: u8, subopcode: u8) -> Option<InstructionMeta>

Parses a sized instruction in form 2.

This covers the opcode range from 0x00 to 0xBF. Form 2 essentially means that b decides the form and subopcode is used to match the actual instruction.

pub fn parse_unsized_form_1(a: u8, b: u8) -> Option<InstructionMeta>

Parses an unsized instruction in form 1.

This covers the opcode range from 0xC0 to 0xFF. Form 1 essentially means that a decides the form and b selects the subopcode.

pub fn parse_unsized_form_2(b: u8, subopcode: u8) -> Option<InstructionMeta>

Parses an unsized instruction in form 2.

This covers the opcode range from 0xC0 to 0xFF. Form 2 essentially means that b decides the form and subopcode is used to match the actual instruction.